refactor to conditionally use the rpc_api module
