+++ src/client/rpc_client.rs
@@ -1,44 +1,23 @@
 // RPC Client
 // ================================================================================================
 //
-#[cfg(not(any(test, feature = "mock")))]
-use crate::errors::RpcApiError;
-use core::fmt;
-#[cfg(not(any(test, feature = "mock")))]
+use crate::{client::RpcApiEndpoint, errors::RpcApiError};
 use miden_node_proto::{
     requests::{SubmitProvenTransactionRequest, SyncStateRequest},
     responses::{SubmitProvenTransactionResponse, SyncStateResponse},
     rpc::api_client::ApiClient,
 };
-#[cfg(not(any(test, feature = "mock")))]
 use tonic::transport::Channel;
 
 // CONSTANTS
 // ================================================================================================
 
-#[derive(Debug)]
-pub enum RpcApiEndpoint {
-    SyncState,
-    SubmitProvenTx,
-}
-
-impl fmt::Display for RpcApiEndpoint {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RpcApiEndpoint::SyncState => write!(f, "sync_state"),
-            RpcApiEndpoint::SubmitProvenTx => write!(f, "submit_proven_transaction"),
-        }
-    }
-}
-
 /// Wrapper for ApiClient which defers establishing a connection with a node until necessary
-#[cfg(not(any(test, feature = "mock")))]
 pub(crate) struct RpcClient {
     rpc_api: Option<ApiClient<Channel>>,
     endpoint: String,
 }
 
-#[cfg(not(any(test, feature = "mock")))]
 impl RpcClient {
     pub fn new(config_endpoint: String) -> RpcClient {
         RpcClient {
